# Add cuDNN implementation for bias_add