@yuanms2 2020-08-16
这个改动会影响bert和cnn吗？ 会让二者更快吗？