@liujuncheng 2020-08-16
> 这个改动会影响bert和cnn吗？ 会让二者更快吗？

针对bert尤其是fp16进行的优化